refactor(callers): migrate yaml-runner imports to engine/executor
Move the six remaining `engine/yaml-runner.js` imports onto
`engine/executor.js` (or `engine/template.js` for the helpers). The
yaml-runner shim stays in place for external users only — every
in-tree caller now points at the new module surface.

- src/cli.ts → executor.js (runPipeline + PipelineError)
- src/protocol/acp.ts → executor.js
- src/commands/health.ts → executor.js
- src/commands/skills.ts → executor.js
- src/mcp/server.ts → executor.js
- src/transport/adapters/http.ts → executor.js (assertSafeRequestUrl)
